requirements.txt
aniso8601==9.0.1
attrs==22.1.0
bcrypt==4.2.1
blinker==1.5
click==8.1.3
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
Flask==2.2.2
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
Flask-WTF==1.2.2
idna==3.10
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
mongomock==4.1.2
ordered-set==4.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.2.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==1.0.1
pytz==2022.2.1
rich==13.9.4
sentinels==1.0.0
serverless-wsgi==3.0.5
six==1.16.0
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.2.2
wrapt==1.17.2
WTForms==3.2.1
zipp==3.8.1
